pyproject.toml
[project]
name = "jeremy-card-pdfs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "faker>=37.8.0",
    "jinja2>=3.1.6",
    "path-py>=12.5.0",
    "pillow>=11.3.0",
    "pyaml>=25.7.0",
    "reportlab>=4.4.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
